pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "reportlab>=4.3.1",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "flask-mail>=0.10.0",
    "sendgrid>=6.11.0",
    "jinja2>=3.1.6",
    "markupsafe>=3.0.2",
    "flask-wtf>=1.2.2",
    "pymysql>=1.1.1",
    "requests>=2.32.3",
]
